all is done

+++ src/Utils/LocalStorageWork.jsx
@@ -26,8 +26,14 @@ export const saveToLocalStorage = (blog) => {
 //delete bookmarks
 
 export const deleteBlogFromLocalStorage = (id) => {
-    let blogs = getDataFromLocalStorage();
-    const newBlogs = blogs.filter(b => b.id!=id);
-    localStorage.setItem('blogs', JSON.stringify(newBlogs));
-    toast.success('Blog removed from bookmark!');
-}
+  let blogs = getDataFromLocalStorage();
+  const isExists = blogs.find((b) => b.id == id);
+  if (isExists) {
+    const newBlogs = blogs.filter((b) => b.id != id);
+    localStorage.setItem("blogs", JSON.stringify(newBlogs));
+    toast.success("Blog removed from bookmark!");
+  }
+  else{
+    toast.error("Bookmark isn't exists!")
+  }
+};
